+++ MaterialDesignExtensions/Controls/OpenMultipleDirectoriesControl.cs
@@ -28,6 +28,7 @@ namespace MaterialDesignExtensions.Controls
     public class OpenMultipleDirectoriesControl : FileSystemControl
     {
         private const string SelectionItemsControlName = "selectionItemsControl";
+        private const string EmptySelectionTextBlockName = "emptySelectionTextBlock";
 
         /// <summary>
         /// Internal command used by the XAML template (public to be available in the XAML template). Not intended for external usage.
@@ -89,6 +90,7 @@ namespace MaterialDesignExtensions.Controls
         }
 
         private ItemsControl m_selectionItemsControl;
+        private TextBlock m_emptySelectionTextBlock;
 
         static OpenMultipleDirectoriesControl()
         {
@@ -101,11 +103,13 @@ namespace MaterialDesignExtensions.Controls
         public OpenMultipleDirectoriesControl()
             : base()
         {
+
+            m_selectionItemsControl = null;
+            m_emptySelectionTextBlock = null;
+
             CommandBindings.Add(new CommandBinding(OpenSelectionDrawerCommand, OpenSelectionDrawerCommandHandler));
             CommandBindings.Add(new CommandBinding(SelectDirectoryCommand, SelectDirectoryCommandHandler));
             CommandBindings.Add(new CommandBinding(SelectMultipleDirectoriesCommand, SelectMultipleDirectoriesCommandHandler, CanExecuteSelectMultipleDirectoriesCommand));
-
-            m_selectionItemsControl = null;
         }
 
         public override void OnApplyTemplate()
@@ -114,7 +118,11 @@ namespace MaterialDesignExtensions.Controls
 
             m_selectionItemsControl = Template.FindName(SelectionItemsControlName, this) as ItemsControl;
 
+            m_emptySelectionTextBlock = Template.FindName(EmptySelectionTextBlockName, this) as TextBlock;
+
             UpdateSelectionList();
+
+            UpdateSelectionListVisibility();
         }
 
         private void OpenSelectionDrawerCommandHandler(object sender, ExecutedRoutedEventArgs args)
@@ -169,6 +177,8 @@ namespace MaterialDesignExtensions.Controls
                 {
                     UpdateSelection();
                     UpdateSelectionList();
+
+                    UpdateSelectionListVisibility();
                 }
             }
 
@@ -213,6 +223,25 @@ namespace MaterialDesignExtensions.Controls
             m_selectionItemsControl.ItemsSource = m_controller.SelectedDirectories
                 .OrderBy(directory => directory.Name.ToLower())
                 .ThenBy(directory => directory.Parent.FullName.ToLower());
+        }
+
+        /// <summary>
+        /// Shows the list of the selected directories or hides it if it is empty. A message will be show instead of an empty list.
+        /// </summary>
+        protected void UpdateSelectionListVisibility()
+        {
+            if (m_selectionItemsControl != null
+                && m_selectionItemsControl.ItemsSource != null
+                && m_selectionItemsControl.ItemsSource.GetEnumerator().MoveNext())
+            {
+                m_selectionItemsControl.Visibility = Visibility.Visible;
+                m_emptySelectionTextBlock.Visibility = Visibility.Collapsed;
+            }
+            else
+            {
+                m_selectionItemsControl.Visibility = Visibility.Collapsed;
+                m_emptySelectionTextBlock.Visibility = Visibility.Visible;
+            }
         }
     }