+++ ReleaseNotes.md
@@ -124,6 +124,7 @@ Please change your configuration according to [App.xaml](https://github.com/spie
 * New window class `MaterialWindow` for a Material Design like styled window
 * Support of `TabStripPlacement` in `TabControl` styles
 * `OpenMultipleDirectoriesControl` and `OpenMultipleDirectoriesDialog` to select multiple directories
+* `OpenMultipleFilesControl` and `OpenMultipleFilesDialog` to select multiple files
 #### Fixes
 * Fixed content layout of horizontal steppers
 #### Obsolete